requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
jinja2>=3.1.0
requests>=2.31.0
openpyxl>=3.1.0
python-docx>=1.0.0
pytest>=7.4.0
